R programming exercise - using Shiny
Read both datasets into R
Set up your R script to work with Shiny. The UI and Server functions do not have to be in separate files.
Create a proportional symbol map and faceted line chart with following specifications and place them side by side.
Create a proportional symbol map using Leaflet that shows the locations of all sensors across Melbourne’s CBD:
Using circle markers, the radius of each sensor’s marker should be proportional to the average hourly count over the entire year of 2019. Use an appropriate scaling factor to ensure no marker is excessively large.
The name of the sensor should be visible in a tooltip when hovered over by mouse.
The output map should be appropriately sized and scaled in the Shiny application to comfortably show all sensors by default.
Create a faceted line chart using ggplot2 that shows the average pedestrian counts throughout each day of the week for the chosen sensor:
The x axis should show the time of day at hourly intervals, that is from 0 (12 am) to 23 (11pm).
The y axis should show the average pedestrian count at the given hour.
The line chart should be faceted by day of the week (i.e., Monday, Tuesday, Wednesday, etc.), and should be ordered appropriately.
Create a Shiny input selection box that includes an option each sensor.
The line chart should update based on the sensor that is chosen in the input selection box.
Create interactive linking between the map and line chart. When a marker on the map is clicked, the input selection box should be updated to the relevant sensor and the line chart accordingly.